make x and y be an i32 in position and add player controller

+++ src/domain/commons/position.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, Clone, Copy)]
 pub struct Position {
-    pub x: usize,
-    pub y: usize
+    pub x: i32,
+    pub y: i32,
 }
 
 impl Position {
-    pub fn new(x: usize, y: usize) -> Self {
+    pub fn new(x: i32, y: i32) -> Self {
         Position {
             x,
             y
